\name{DFPmcmc-package}
\alias{DFPmcmc-package}
\alias{DFPmcmc}
\docType{package}
\title{
\packageTitle{DFPmcmc}
}
\description{
\packageDescription{DFPmcmc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DFPmcmc}
\packageIndices{DFPmcmc}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{DFPmcmc}

Maintainer: \packageMaintainer{DFPmcmc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
